requirements.txt
mkdocs
mkdocs-material
mkdocs-include-dir-to-nav
mkdocs-markdownextradata-plugin
pymdown-extensions
mkdocs-macros-plugin
mkdocs-rss-plugin
mkdocs-exclude-search
mkdocs-git-revision-date-localized-plugin
mkdocs-git-authors-plugin
mkdocs-open-in-new-tab
mkdocs-glightbox
pyyaml
